ffai/requirements.txt
pandas>=2.1.0
numpy>=1.24.0
torch>=2.0.0
espn_api
requests>=2.28.0
colorlog
matplotlib>=3.8.0
seaborn>=0.13.0
scikit-learn>=1.3.0
pyyaml>=6.0.0
tqdm>=4.65.0
playwright>=1.40.0
